requirements.txt
torch
datasets
transformers
pre-commit
huggingface_hub
python-dotenv